fix: Feed 컴포넌트 구조 변경 - data를 props로 받음
